cdn/requirements.txt
django
psycopg2-binary
djangorestframework
pillow
django-cors-headers
requests